Fix: fire splitTransition events in a new transaction.
Current implementation breaks Airstream's contract on transactions (they certainly should not be null), and actually caused a bug in one user's code derived from zio-petclinic (which uses splitTransition) - Airstream basically stopped working after one splitTransition event. The bug itself was apparently dependent on specific data (arrangement of items provided to splitTransition), and was hard to minimize, but this fix eliminated it, and I believe this null transaction is likely the root cause.